Split out nodatime, and systemtextjson from the core package, so they can be used independently.
