+++ Packages.props
@@ -32,6 +32,7 @@
     <PackageReference Update="NodaTime.Serialization.SystemTextJson" Version="1.0.0" />
     <PackageReference Update="Rocket.Surgery.Conventions" Version="10.0.0-beta.6" />
     <PackageReference Update="System.Text.Json" Version="4.7.2" />
+    <PackageReference Update="Newtonsoft.Json" Version="12.0.3" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Update="Bogus" Version="29.0.2" />
