+++ src/Extensions.AutoMapper.NewtonsoftJson/JsonDefaultValue.cs
@@ -1,0 +1,7 @@
+﻿namespace Rocket.Surgery.Extensions.AutoMapper.NewtonsoftJson {
+    internal enum JsonDefaultValue
+    {
+        Default,
+        NotNull
+    }
+}
