+++ src/Extensions.AutoMapper.NodaTime/Converters/LocalTimeConverter.cs
@@ -3,29 +3,29 @@ using AutoMapper;
 using JetBrains.Annotations;
 using NodaTime;
 
-namespace Rocket.Surgery.Extensions.AutoMapper.Converters
+namespace Rocket.Surgery.Extensions.AutoMapper.NodaTime.Converters
 {
     /// <summary>
-    /// InstantConverter.
+    /// LocalTimeConverter.
     /// Implements the <see cref="ITypeConverter{TSource,TDestination}" />
-    /// Implements the <see cref="ITypeConverter{Instant, DateTimeOffset}" />
-    /// Implements the <see cref="ITypeConverter{DateTime, Instant}" />
-    /// Implements the <see cref="ITypeConverter{DateTimeOffset, Instant}" />
+    /// Implements the <see cref="ITypeConverter{TimeSpan, LocalTime}" />
+    /// Implements the <see cref="ITypeConverter{LocalTime, DateTime}" />
+    /// Implements the <see cref="ITypeConverter{DateTime, LocalTime}" />
     /// </summary>
-    /// <seealso cref="ITypeConverter{Instant, DateTime}" />
-    /// <seealso cref="ITypeConverter{Instant, DateTimeOffset}" />
-    /// <seealso cref="ITypeConverter{DateTime, Instant}" />
-    /// <seealso cref="ITypeConverter{DateTimeOffset, Instant}" />
+    /// <seealso cref="ITypeConverter{LocalTime, TimeSpan}" />
+    /// <seealso cref="ITypeConverter{TimeSpan, LocalTime}" />
+    /// <seealso cref="ITypeConverter{LocalTime, DateTime}" />
+    /// <seealso cref="ITypeConverter{DateTime, LocalTime}" />
     [PublicAPI]
-    public class InstantConverter :
-        ITypeConverter<Instant, DateTime>,
-        ITypeConverter<Instant?, DateTime?>,
-        ITypeConverter<Instant, DateTimeOffset>,
-        ITypeConverter<Instant?, DateTimeOffset?>,
-        ITypeConverter<DateTime, Instant>,
-        ITypeConverter<DateTime?, Instant?>,
-        ITypeConverter<DateTimeOffset, Instant>,
-        ITypeConverter<DateTimeOffset?, Instant?>
+    public class LocalTimeConverter :
+        ITypeConverter<LocalTime, TimeSpan>,
+        ITypeConverter<LocalTime?, TimeSpan?>,
+        ITypeConverter<TimeSpan, LocalTime>,
+        ITypeConverter<TimeSpan?, LocalTime?>,
+        ITypeConverter<LocalTime, DateTime>,
+        ITypeConverter<LocalTime?, DateTime?>,
+        ITypeConverter<DateTime, LocalTime>,
+        ITypeConverter<DateTime?, LocalTime?>
     {
         /// <summary>
         /// Performs conversion from source to destination type
@@ -34,14 +34,8 @@ namespace Rocket.Surgery.Extensions.AutoMapper.Converters
         /// <param name="destination">Destination object</param>
         /// <param name="context">Resolution context</param>
         /// <returns>Destination object</returns>
-        public Instant Convert(DateTime source, Instant destination, ResolutionContext context)
-        {
-            var utcDateTime = source.Kind == DateTimeKind.Unspecified
-                ? DateTime.SpecifyKind(source, DateTimeKind.Utc)
-                : source.ToUniversalTime();
-
-            return Instant.FromDateTimeUtc(utcDateTime);
-        }
+        public LocalTime Convert(DateTime source, LocalTime destination, ResolutionContext context)
+            => LocalDateTime.FromDateTime(source).TimeOfDay;
 
         /// <summary>
         /// Performs conversion from source to destination type
@@ -50,19 +44,8 @@ namespace Rocket.Surgery.Extensions.AutoMapper.Converters
         /// <param name="destination">Destination object</param>
         /// <param name="context">Resolution context</param>
         /// <returns>Destination object</returns>
-        public Instant? Convert(DateTime? source, Instant? destination, ResolutionContext context)
-        {
-            if (!source.HasValue)
-            {
-                return destination;
-            }
-
-            var utcDateTime = source.Value.Kind == DateTimeKind.Unspecified
-                ? DateTime.SpecifyKind(source.Value, DateTimeKind.Utc)
-                : source.Value.ToUniversalTime();
-
-            return Instant.FromDateTimeUtc(utcDateTime);
-        }
+        public LocalTime? Convert(DateTime? source, LocalTime? destination, ResolutionContext context)
+            => source.HasValue ? LocalDateTime.FromDateTime(source.Value).TimeOfDay : destination;
 
         /// <summary>
         /// Performs conversion from source to destination type
@@ -71,8 +54,8 @@ namespace Rocket.Surgery.Extensions.AutoMapper.Converters
         /// <param name="destination">Destination object</param>
         /// <param name="context">Resolution context</param>
         /// <returns>Destination object</returns>
-        public Instant Convert(DateTimeOffset source, Instant destination, ResolutionContext context)
-            => Instant.FromDateTimeOffset(source);
+        public DateTime Convert(LocalTime source, DateTime destination, ResolutionContext context)
+            => source.On(new LocalDate(1, 1, 1)).ToDateTimeUnspecified();
 
         /// <summary>
         /// Performs conversion from source to destination type
@@ -81,8 +64,8 @@ namespace Rocket.Surgery.Extensions.AutoMapper.Converters
         /// <param name="destination">Destination object</param>
         /// <param name="context">Resolution context</param>
         /// <returns>Destination object</returns>
-        public Instant? Convert(DateTimeOffset? source, Instant? destination, ResolutionContext context)
-            => source.HasValue ? Instant.FromDateTimeOffset(source.Value) : destination;
+        public TimeSpan Convert(LocalTime source, TimeSpan destination, ResolutionContext context)
+            => new TimeSpan(source.TickOfDay);
 
         /// <summary>
         /// Performs conversion from source to destination type
@@ -91,8 +74,8 @@ namespace Rocket.Surgery.Extensions.AutoMapper.Converters
         /// <param name="destination">Destination object</param>
         /// <param name="context">Resolution context</param>
         /// <returns>Destination object</returns>
-        public DateTime Convert(Instant source, DateTime destination, ResolutionContext context)
-            => source.ToDateTimeUtc();
+        public DateTime? Convert(LocalTime? source, DateTime? destination, ResolutionContext context)
+            => source?.On(new LocalDate(1, 1, 1)).ToDateTimeUnspecified() ?? destination;
 
         /// <summary>
         /// Performs conversion from source to destination type
@@ -101,8 +84,8 @@ namespace Rocket.Surgery.Extensions.AutoMapper.Converters
         /// <param name="destination">Destination object</param>
         /// <param name="context">Resolution context</param>
         /// <returns>Destination object</returns>
-        public DateTimeOffset Convert(Instant source, DateTimeOffset destination, ResolutionContext context)
-            => source.ToDateTimeOffset();
+        public TimeSpan? Convert(LocalTime? source, TimeSpan? destination, ResolutionContext context)
+            => source.HasValue ? new TimeSpan(source.Value.TickOfDay) : destination;
 
         /// <summary>
         /// Performs conversion from source to destination type
@@ -111,8 +94,8 @@ namespace Rocket.Surgery.Extensions.AutoMapper.Converters
         /// <param name="destination">Destination object</param>
         /// <param name="context">Resolution context</param>
         /// <returns>Destination object</returns>
-        public DateTime? Convert(Instant? source, DateTime? destination, ResolutionContext context)
-            => source?.ToDateTimeUtc() ?? destination;
+        public LocalTime Convert(TimeSpan source, LocalTime destination, ResolutionContext context)
+            => LocalTime.FromTicksSinceMidnight(source.Ticks);
 
         /// <summary>
         /// Performs conversion from source to destination type
@@ -121,7 +104,7 @@ namespace Rocket.Surgery.Extensions.AutoMapper.Converters
         /// <param name="destination">Destination object</param>
         /// <param name="context">Resolution context</param>
         /// <returns>Destination object</returns>
-        public DateTimeOffset? Convert(Instant? source, DateTimeOffset? destination, ResolutionContext context)
-            => source?.ToDateTimeOffset() ?? destination;
+        public LocalTime? Convert(TimeSpan? source, LocalTime? destination, ResolutionContext context)
+            => source.HasValue ? LocalTime.FromTicksSinceMidnight(source.Value.Ticks) : destination;
     }
 }
