+++ src/Extensions.AutoMapper.NodaTime/Converters/OffsetConverter.cs
@@ -3,21 +3,21 @@ using AutoMapper;
 using JetBrains.Annotations;
 using NodaTime;
 
-namespace Rocket.Surgery.Extensions.AutoMapper.Converters
+namespace Rocket.Surgery.Extensions.AutoMapper.NodaTime.Converters
 {
     /// <summary>
-    /// OffsetDateTimeConverter.
+    /// OffsetConverter.
     /// Implements the <see cref="ITypeConverter{TSource,TDestination}" />
-    /// Implements the <see cref="ITypeConverter{DateTimeOffset, OffsetDateTime}" />
+    /// Implements the <see cref="ITypeConverter{TimeSpan, Offset}" />
     /// </summary>
-    /// <seealso cref="ITypeConverter{OffsetDateTime, DateTimeOffset}" />
-    /// <seealso cref="ITypeConverter{DateTimeOffset, OffsetDateTime}" />
+    /// <seealso cref="ITypeConverter{Offset, TimeSpan}" />
+    /// <seealso cref="ITypeConverter{TimeSpan, Offset}" />
     [PublicAPI]
-    public class OffsetDateTimeConverter :
-        ITypeConverter<OffsetDateTime, DateTimeOffset>,
-        ITypeConverter<OffsetDateTime?, DateTimeOffset?>,
-        ITypeConverter<DateTimeOffset, OffsetDateTime>,
-        ITypeConverter<DateTimeOffset?, OffsetDateTime?>
+    public class OffsetConverter :
+        ITypeConverter<Offset, TimeSpan>,
+        ITypeConverter<Offset?, TimeSpan?>,
+        ITypeConverter<TimeSpan, Offset>,
+        ITypeConverter<TimeSpan?, Offset?>
     {
         /// <summary>
         /// Performs conversion from source to destination type
@@ -26,8 +26,7 @@ namespace Rocket.Surgery.Extensions.AutoMapper.Converters
         /// <param name="destination">Destination object</param>
         /// <param name="context">Resolution context</param>
         /// <returns>Destination object</returns>
-        public OffsetDateTime Convert(DateTimeOffset source, OffsetDateTime destination, ResolutionContext context)
-            => OffsetDateTime.FromDateTimeOffset(source);
+        public TimeSpan Convert(Offset source, TimeSpan destination, ResolutionContext context) => source.ToTimeSpan();
 
         /// <summary>
         /// Performs conversion from source to destination type
@@ -36,8 +35,8 @@ namespace Rocket.Surgery.Extensions.AutoMapper.Converters
         /// <param name="destination">Destination object</param>
         /// <param name="context">Resolution context</param>
         /// <returns>Destination object</returns>
-        public OffsetDateTime? Convert(DateTimeOffset? source, OffsetDateTime? destination, ResolutionContext context)
-            => source.HasValue ? OffsetDateTime.FromDateTimeOffset(source.Value) : destination;
+        public TimeSpan? Convert(Offset? source, TimeSpan? destination, ResolutionContext context)
+            => source?.ToTimeSpan() ?? destination;
 
         /// <summary>
         /// Performs conversion from source to destination type
@@ -46,8 +45,8 @@ namespace Rocket.Surgery.Extensions.AutoMapper.Converters
         /// <param name="destination">Destination object</param>
         /// <param name="context">Resolution context</param>
         /// <returns>Destination object</returns>
-        public DateTimeOffset Convert(OffsetDateTime source, DateTimeOffset destination, ResolutionContext context)
-            => source.ToDateTimeOffset();
+        public Offset Convert(TimeSpan source, Offset destination, ResolutionContext context)
+            => Offset.FromTicks(source.Ticks);
 
         /// <summary>
         /// Performs conversion from source to destination type
@@ -56,7 +55,7 @@ namespace Rocket.Surgery.Extensions.AutoMapper.Converters
         /// <param name="destination">Destination object</param>
         /// <param name="context">Resolution context</param>
         /// <returns>Destination object</returns>
-        public DateTimeOffset? Convert(OffsetDateTime? source, DateTimeOffset? destination, ResolutionContext context)
-            => source?.ToDateTimeOffset() ?? destination;
+        public Offset? Convert(TimeSpan? source, Offset? destination, ResolutionContext context)
+            => source.HasValue ? Offset.FromTicks(source.Value.Ticks) : destination;
     }
 }
