+++ src/Extensions.AutoMapper.NodaTime/Rocket.Surgery.Extensions.AutoMapper.NodaTime.csproj
@@ -5,7 +5,6 @@
     </PropertyGroup>
     <ItemGroup>
         <PackageReference Include="AutoMapper" />
-        <PackageReference Include="System.Text.Json" />
-        <PackageReference Include="Newtonsoft.Json" />
+        <PackageReference Include="NodaTime" />
     </ItemGroup>
 </Project>
