+++ src/Extensions.AutoMapper.SystemTextJson/JsonDefaultValue.cs
@@ -1,0 +1,7 @@
+﻿namespace Rocket.Surgery.Extensions.AutoMapper.SystemTextJson {
+    internal enum JsonDefaultValue
+    {
+        Default,
+        NotNull
+    }
+}
