--- src/Extensions.AutoMapper/JsonDefaultValue.cs
@@ -1,7 +1,0 @@
-﻿namespace Rocket.Surgery.Extensions.AutoMapper {
-    public enum JsonDefaultValue
-    {
-        Default,
-        NotNull
-    }
-}
