+++ test/AutoMapper.Tests/OffsetDateTimeTests.cs
@@ -4,7 +4,7 @@ using System.Reflection;
 using AutoMapper;
 using FluentAssertions;
 using NodaTime;
-using Rocket.Surgery.Extensions.AutoMapper.Converters;
+using Rocket.Surgery.Extensions.AutoMapper.NodaTime.Converters;
 using Xunit;
 using Xunit.Abstractions;
