+++ test/AutoMapper.Tests/TypeConverterTest.cs
@@ -10,6 +10,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using NodaTime;
 using Rocket.Surgery.Conventions.AutoMapper;
+using Rocket.Surgery.Extensions.AutoMapper.NodaTime;
 using Rocket.Surgery.Extensions.Testing;
 using Serilog.Events;
 using Xunit.Abstractions;
